% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot_density_sf.R
\name{dot_density_sf}
\alias{dot_density_sf}
\title{dot_density_sf}
\usage{
dot_density_sf(
  sf = NULL,
  dot_col = NULL,
  cases_per_dot = NULL,
  group_col = NULL
)
}
\arguments{
\item{sf}{A simple features object of class "sf".}

\item{dot_col}{The variabe name from \code{sf} that represents the dots}

\item{cases_per_dot}{A numeric that sets the number of cases represented by each dot}

\item{group_col}{The variabe name from \code{sf} that represents the cases into distinct groups}
}
\value{
A \code{sf} object with geometry type \code{POINT}. If the
the cases are grouped, then a feature is added to \code{sf} that represents the fraction of dots for each group
}
\description{
The function returns a special feature object (class \code{sf}) from the submitted simple feature object
that has \code{POINT} geometries (called a "dot")
representing each observation (or counts of observations) of the feature object.
}
\details{
The functions \code{terra::vect()} and \code{terra::dots()} are used in creating the dots
and can be used to make a dot-density map.

See \href{https://rspatial.github.io/terra/reference/dots.html}{Make a dot-density map}
for more information and example on dot-density maps.
}
\author{
Rick Dean
}
