% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_to_sf.R
\name{matrix_to_sf}
\alias{matrix_to_sf}
\title{matrix_to_sf}
\usage{
matrix_to_sf(a_matrix = NULL, xy_v = NULL, crs = sf::st_crs(4326))
}
\arguments{
\item{a_matrix}{A matrix describing the point/polygon geometry.}

\item{xy_v}{A two element vector that defines the columns of \code{a_matrix} for the x/y coordinates.
Values can be the column names of the matrix or column index numbers.}

\item{crs}{Defines the coordinate reference system to use. Can be an integer with the EPSG code or character with proj4 string.
Default is EPSG value 4326.}
}
\value{
An sf::sf object
}
\description{
Function converts a matrix of long/lat coordinate(s) into a simple features
object (sf::sf)
}
\details{
The submitted matrix of longitude (column 1) and latitude (column 2)
coordinate(s) is converted into an sf object. See \href{https://r-spatial.github.io/sf/articles/sf1.html}{Simple Features for R}
for more information on simple features.
}
\author{
Rick Dean
}
