% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfc_to_matrix.R
\name{sfc_to_matrix}
\alias{sfc_to_matrix}
\title{sfc_to_matrix}
\usage{
sfc_to_matrix(x)
}
\arguments{
\item{x}{An 'sfc' object}
}
\value{
A numeric matrix of the point coordinates from an sf::sfc object.
}
\description{
Function converts point geometries from an sf::sfc object to matrix form
}
\author{
Rick Dean
}
